+++ src/edit/Caret.ts
@@ -61,8 +61,22 @@ import TypeVariable from '../nodes/TypeVariable';
 import UnicodeString from '../unicode/UnicodeString';
 
 export type InsertionContext = { before: Node[]; after: Node[] };
-export type CaretPosition = number | Node;
+
+/**
+ * Selections can be a single text buffer position, a range, or a node.
+ */
+export type CaretPosition = number | Node | [number, number];
 export type Entry = 'previous' | 'next' | undefined;
+
+export function isPosition(position: CaretPosition): position is number {
+    return typeof position === 'number';
+}
+export function isRange(position: CaretPosition): position is [number, number] {
+    return Array.isArray(position);
+}
+export function isNode(position: CaretPosition): position is Node {
+    return position instanceof Node;
+}
 
 export default class Caret {
     readonly source: Source;
@@ -129,11 +143,13 @@ export default class Caret {
     }
 
     getNodeInside() {
-        return typeof this.position === 'number'
+        return isPosition(this.position)
             ? this.insideToken()
                 ? this.tokenExcludingSpace
                 : undefined
-            : this.position;
+            : isNode(this.position)
+              ? this.position
+              : undefined;
     }
 
     getTokenPrior() {
@@ -159,10 +175,7 @@ export default class Caret {
     }
 
     insideToken() {
-        if (
-            this.tokenExcludingSpace === undefined ||
-            this.position instanceof Node
-        )
+        if (this.tokenExcludingSpace === undefined || !this.isPosition())
             return false;
         const start = this.source.getTokenTextPosition(
             this.tokenExcludingSpace,
@@ -267,10 +280,7 @@ export default class Caret {
         const empty = { before: [], after: [] };
 
         // If the caret is a node, there is no notion of between.
-        if (
-            this.position instanceof Node ||
-            this.tokenIncludingSpace === undefined
-        )
+        if (!this.isPosition() || this.tokenIncludingSpace === undefined)
             return empty;
 
         // Get the line number of the position.
@@ -415,27 +425,32 @@ export default class Caret {
         return typeof this.position === 'number';
     }
 
+    isRange(): this is { position: [number, number] } {
+        return Array.isArray(this.position);
+    }
+
     isAtPropertyReference() {
-        if (this.position instanceof Node) return false;
+        if (!this.isPosition()) return false;
         return this.source.getCode().at(this.position - 1) === PROPERTY_SYMBOL;
     }
 
     /** True if this caret's position is or is inside of the given node. */
     isIn(node: Node, includeEnd: boolean) {
-        if (this.position instanceof Node)
+        if (this.isNode())
             return (
                 this.position === node ||
                 this.source.root.hasAncestor(this.position, node)
             );
-
-        const start = this.source.getNodeFirstPosition(node);
-        const end = this.source.getNodeLastPosition(node);
-        return (
-            start !== undefined &&
-            end !== undefined &&
-            start <= this.position &&
-            (includeEnd ? this.position <= end : this.position < end)
-        );
+        if (this.isPosition()) {
+            const start = this.source.getNodeFirstPosition(node);
+            const end = this.source.getNodeLastPosition(node);
+            return (
+                start !== undefined &&
+                end !== undefined &&
+                start <= this.position &&
+                (includeEnd ? this.position <= end : this.position < end)
+            );
+        } else return false;
     }
 
     /** Get the code position corresponding to the beginning of the given row.  */
@@ -469,7 +484,7 @@ export default class Caret {
     moveInline(sibling: boolean, direction: -1 | 1): Caret {
         // Map the direction onto an entry direction.
         const entry = direction > 0 ? 'next' : 'previous';
-        if (this.position instanceof Node) {
+        if (this.isNode()) {
             // If sibling, then find the parent of the current node and choose it's sibling.
             // If there's no sibling in this direction, then do nothing.
             if (sibling) {
@@ -516,7 +531,12 @@ export default class Caret {
                     ? this.withPosition(index, this.column, entry)
                     : this;
             }
-        } else {
+        } else if (this.isPosition() || this.isRange()) {
+            const currentPosition =
+                typeof this.position === 'number'
+                    ? this.position
+                    : this.position[1];
+
             // At the beginning or end? Do nothing.
             if (
                 this.position ===
@@ -525,9 +545,9 @@ export default class Caret {
                 return this;
 
             // At a token start and going next? Select the token.
-            const token = this.source.getTokenAt(this.position, false);
+            const token = this.source.getTokenAt(currentPosition, false);
             const tokenBefore = this.source.getTokenAt(
-                this.position - 1,
+                currentPosition - 1,
                 false,
             );
 
@@ -555,11 +575,13 @@ export default class Caret {
                 );
 
             return this.withPosition(
-                this.position + direction,
-                this.source.getColumn(this.position + direction),
+                currentPosition + direction,
+                this.source.getColumn(currentPosition + direction),
                 entry,
             );
         }
+        // Some other mystery selection type? Do nothing.
+        else return this;
     }
 
     static isBlockEditable(token: Token): boolean {
@@ -731,12 +753,13 @@ export default class Caret {
     /** Move to the next node or position in blocks mode. */
     moveInlineSemantic(direction: -1 | 1): Caret | undefined {
         // Find the current position.
-        const currentPosition =
-            typeof this.position === 'number'
-                ? this.position
-                : direction < 0
-                  ? this.source.getNodeFirstPosition(this.position)
-                  : this.source.getNodeLastPosition(this.position);
+        const currentPosition = isPosition(this.position)
+            ? this.position
+            : isRange(this.position)
+              ? this.position[0]
+              : direction < 0
+                ? this.source.getNodeFirstPosition(this.position)
+                : this.source.getNodeLastPosition(this.position);
 
         // No current position for some reason? No change;
         if (currentPosition === undefined) return undefined;
@@ -774,11 +797,64 @@ export default class Caret {
         return undefined;
     }
 
+    expandInline(amount: number) {
+        // Currently a position? Create a range with the current position as the start, and the end with the adjustment.
+        if (isPosition(this.position)) {
+            return this.withPosition([
+                this.position,
+                Math.max(
+                    0,
+                    Math.min(
+                        this.position + amount,
+                        this.source.getCode().getLength(),
+                    ),
+                ),
+            ]);
+        }
+        // Already a range? Expand the end.
+        else if (this.isRange()) {
+            const [start, end] = this.position;
+            return this.withPosition([
+                start,
+                Math.max(
+                    0,
+                    Math.min(end + amount, this.source.getCode().getLength()),
+                ),
+            ]);
+        }
+        // Node? Don't do anything.
+        else return this;
+    }
+
+    expandVertically(amount: -1 | 1) {
+        if (isPosition(this.position)) {
+            const verticalPosition = this.getVertical(amount, this.position);
+            if (verticalPosition && verticalPosition.isPosition())
+                return this.withPosition([
+                    this.position,
+                    verticalPosition.position,
+                ]);
+            else return this;
+        }
+        // Already a range? Expand the end.
+        else if (this.isRange()) {
+            const [start, end] = this.position;
+            const verticalPosition = this.getVertical(amount, end);
+            if (verticalPosition && verticalPosition.isPosition())
+                return this.withPosition([start, verticalPosition.position]);
+            else return this;
+        }
+        // Node? Don't do anything.
+        else return this;
+    }
+
     atLineBoundary(start: boolean): Caret {
-        let position =
-            this.position instanceof Node
-                ? this.source.getNodeFirstPosition(this.position)
-                : this.position;
+        let position = isNode(this.position)
+            ? this.source.getNodeFirstPosition(this.position)
+            : isPosition(this.position)
+              ? this.position
+              : this.position[0];
+
         if (position === undefined) return this;
         if (start) {
             if (this.source.code.at(position - 1) === '\n') return this;
@@ -816,7 +892,7 @@ export default class Caret {
     }
 
     getTextPosition(start: boolean, offset = 0): number | undefined {
-        if (this.position instanceof Node) {
+        if (isNode(this.position)) {
             // Get the first or last token of the given node.
             const tokens = this.position.nodes(
                 (n): n is Token => n instanceof Token,
@@ -832,7 +908,8 @@ export default class Caret {
                     return index + last.getTextLength() + offset;
             }
             return undefined;
-        } else return this.position;
+        } else if (isPosition(this.position)) return this.position;
+        else return this.position[0];
     }
 
     getPlaceholderAtPosition(position: number): Node | undefined {
@@ -848,7 +925,7 @@ export default class Caret {
     }
 
     withPosition(
-        position: number | Node,
+        position: CaretPosition,
         column?: number,
         entry?: Entry,
     ): Caret {
@@ -910,7 +987,7 @@ export default class Caret {
     }
 
     insertNode(node: Node, offset: number): Edit | undefined {
-        if (this.position instanceof Node) {
+        if (isNode(this.position)) {
             const position = this.source.getNodeFirstPosition(this.position);
             if (position === undefined) return undefined;
 
@@ -920,6 +997,28 @@ export default class Caret {
                 new Caret(
                     newSource,
                     position + offset,
+                    this.column,
+                    undefined,
+                    node,
+                ),
+            ];
+        } else if (isRange(this.position)) {
+            let newSource: Source | undefined = this.source;
+            const [start, end] = this.position;
+            // Without selection
+            newSource = newSource.withoutGraphemesBetween(start, end);
+            if (newSource === undefined) return undefined;
+            // With node text
+            newSource = newSource.withGraphemesAt(
+                node.toWordplay(getPreferredSpaces(node)),
+                start,
+            );
+            if (newSource === undefined) return undefined;
+            return [
+                newSource,
+                new Caret(
+                    newSource,
+                    start + offset,
                     this.column,
                     undefined,
                     node,
@@ -972,7 +1071,7 @@ export default class Caret {
         if (this.insertionCompletesDelimiter(text))
             return [this.source, this.withPosition(this.position + 1)];
 
-        // Before doing insertion, see if a node is selected, and if wrap or remove it.
+        // Before doing insertion, see if a node is selected, and if so, wrap or remove it.
         if (this.position instanceof Node) {
             // Try wrapping the node
             const wrap = this.wrap(text);
@@ -983,16 +1082,33 @@ export default class Caret {
             // If that didn't do anything, do nothing; it's not removeable.
             if (edit === undefined) return;
             const [source, caret] = edit;
-            if (caret.position instanceof Node) return;
+            if (!isPosition(caret.position)) return;
 
             // Otherwise, we deleted it! Update the source and position.
             newSource = source;
             newPosition = caret.position;
             originalPosition = caret.position;
-        } else {
+        }
+        // If it's a single position, do nothing.
+        else if (isPosition(this.position)) {
             newSource = this.source;
             newPosition = this.position;
             originalPosition = this.position;
+        }
+        // If it's a range, delete the range.
+        else {
+            const [start, end] = this.position;
+            if (start === end) {
+                newSource = this.source;
+                newPosition = start;
+                originalPosition = start;
+            } else {
+                const edit = this.source.withoutGraphemesBetween(start, end);
+                if (edit === undefined) return;
+                newSource = edit;
+                newPosition = start;
+                originalPosition = start;
+            }
         }
 
         // Let's see if we should do any delimiter completion before we insert.
@@ -1445,7 +1561,7 @@ export default class Caret {
         const offset = forward ? 0 : -1;
 
         // If the position is a number, see if this is a rename
-        if (typeof this.position === 'number') {
+        if (isPosition(this.position)) {
             // Otherwise, figure out what to delete.
             // Are we in the middle of a name or at it's end?
             const rename = forward
@@ -1594,7 +1710,7 @@ export default class Caret {
             }
         }
         // If it's a node, replace it with a placeholder, or delete it.
-        else {
+        else if (isNode(this.position)) {
             // Get the parent of the node.
             const node = this.position;
             const parent = this.source.root.getParent(node);
@@ -1688,6 +1804,23 @@ export default class Caret {
             }
             // Otherwise, do nothing.
         }
+        // Range? Delete the range.
+        else {
+            const [start, end] = this.position;
+            // No range deletions in valid only unless both positions are in the same token.
+            if (
+                validOnly &&
+                this.source.getTokenAt(start) !== this.source.getTokenAt(end)
+            )
+                return;
+
+            const newSource = this.source.withoutGraphemesBetween(start, end);
+            if (newSource === undefined) return;
+            return [
+                newSource,
+                this.withPosition(start).withAddition(undefined),
+            ];
+        }
     }
 
     deleteNode(
@@ -1741,11 +1874,15 @@ export default class Caret {
     }
 
     moveVertical(direction: 1 | -1): Edit | undefined {
-        if (this.position instanceof Node) {
+        if (isNode(this.position)) {
             const position = this.source.getNodeFirstPosition(this.position);
             if (position === undefined) return;
             return this.getVertical(direction, position);
-        } else return this.getVertical(direction, this.position);
+        } else
+            return this.getVertical(
+                direction,
+                isPosition(this.position) ? this.position : this.position[0],
+            );
     }
 
     getVertical(direction: 1 | -1, position: number): Caret | undefined {
@@ -1797,9 +1934,11 @@ export default class Caret {
         // Find the token we're at
         const token = node
             ? node
-            : typeof this.position === 'number'
+            : isPosition(this.position)
               ? (this.tokenExcludingSpace ?? this.tokenPrior)
-              : this.position;
+              : isRange(this.position)
+                ? undefined
+                : this.position;
 
         if (token === undefined) return;
 
@@ -1853,13 +1992,21 @@ export default class Caret {
         }
 
         /** If the caret is a node, describe the node. */
-        if (this.position instanceof Node) {
+        if (isNode(this.position)) {
             return locales
                 .concretize(
                     (l) => l.ui.edit.node,
                     new NodeRef(this.position, locales, context),
                     type ? new NodeRef(type, locales, context) : undefined,
                 )
+                .toText();
+        }
+
+        /** If the position is a range, describe the start and end token */
+        if (isRange(this.position)) {
+            const [start, end] = this.position;
+            return locales
+                .concretize((l) => l.ui.edit.range, start, end)
                 .toText();
         }