+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 
 - By student request, allow custom characters in profile and creator views.
 - By student request, added Comic Neue as a supported font! Long live Comic Sans.
+- Fixed #264, adding text selection to code editor.
 
 ### Fixed